@@ -1,6 +1,6 @@
 [metadata]
 name = setuptools
-version = 67.8.0
+version = 68.0.0
 author = Python Packaging Authority
 author_email = distutils-sig@python.org
 description = Easily download, build, install, upgrade, and uninstall Python packages
@@ -25,7 +25,7 @@ project_urls =
 packages = find_namespace:
 # disabled as it causes tests to be included #2505
 # include_package_data = true
-python_requires = >=3.7
+python_requires = >=3.8
 install_requires =
 
 [options.packages.find]
@@ -52,7 +52,7 @@ testing =
 	pytest-mypy >= 0.9.1; \
 		# workaround for jaraco/skeleton#22
 		python_implementation != "PyPy"
-	pytest-enabler >= 1.3
+	pytest-enabler >= 2.2
 	# workaround for pypa/setuptools#3921
 	pytest-ruff; sys_platform != "cygwin"
 
@@ -71,6 +71,8 @@ testing =
 	tomli-w>=1.0.0
 	pytest-timeout
 	pytest-perf
+	# for tools/finalize.py
+	jaraco.develop >= 7.21; python_version >= "3.9" and sys_platform != "cygwin"
 
 testing-integration =
 	pytest
@@ -161,7 +163,6 @@ egg_info.writers =
 	eager_resources.txt = setuptools.command.egg_info:overwrite_arg
 	namespace_packages.txt = setuptools.command.egg_info:overwrite_arg
 	top_level.txt = setuptools.command.egg_info:write_toplevel_names
-	depends.txt = setuptools.command.egg_info:warn_depends_obsolete
 	dependency_links.txt = setuptools.command.egg_info:overwrite_arg
 
 [egg_info]
