@@ -8,12 +8,16 @@ pytest_plugins = 'setuptools.tests.fixtures'
 
 def pytest_addoption(parser):
     parser.addoption(
-        "--package_name", action="append", default=[],
+        "--package_name",
+        action="append",
+        default=[],
         help="list of package_name to pass to test functions",
     )
     parser.addoption(
-        "--integration", action="store_true", default=False,
-        help="run integration tests (only)"
+        "--integration",
+        action="store_true",
+        default=False,
+        help="run integration tests (only)",
     )
 
 
@@ -39,6 +43,10 @@ collect_ignore = [
 if sys.version_info < (3, 6):
     collect_ignore.append('docs/conf.py')  # uses f-strings
     collect_ignore.append('pavement.py')
+
+
+if sys.version_info < (3, 9) or sys.platform == 'cygwin':
+    collect_ignore.append('tools/finalize.py')
 
 
 @pytest.fixture(autouse=True)
