@@ -181,7 +181,6 @@ class egg_info(InfoCommon, Command):
         self.egg_name = None
         self.egg_info = None
         self.egg_version = None
-        self.broken_egg_info = False
         self.ignore_egg_info_in_manifest = False
 
     ####################################
@@ -236,8 +235,6 @@ class egg_info(InfoCommon, Command):
         self.egg_info = _normalization.filename_component(self.egg_name) + '.egg-info'
         if self.egg_base != os.curdir:
             self.egg_info = os.path.join(self.egg_base, self.egg_info)
-        if '-' in self.egg_name:
-            self.check_broken_egg_info()
 
         # Set package version for the benefit of dumber commands
         # (e.g. sdist, bdist_wininst, etc.)
@@ -324,25 +321,6 @@ class egg_info(InfoCommon, Command):
         mm.manifest = manifest_filename
         mm.run()
         self.filelist = mm.filelist
-
-    def check_broken_egg_info(self):
-        bei = self.egg_name + '.egg-info'
-        if self.egg_base != os.curdir:
-            bei = os.path.join(self.egg_base, bei)
-        if os.path.exists(bei):
-            EggInfoDeprecationWarning.emit(
-                "Invalid egg-info directory name.",
-                f"""
-                Your current .egg-info directory has a '-' in its name;
-                this will not work correctly with setuptools commands.
-
-                Please rename {bei!r} to {self.egg_info!r} to correct this problem.
-                """,
-                due_date=(2023, 6, 1),
-                # Old warning, introduced in 2005, might be safe to remove soon
-            )
-            self.broken_egg_info = self.egg_info
-            self.egg_info = bei  # make it work for now
 
 
 class FileList(_FileList):
@@ -706,17 +684,13 @@ def write_pkg_info(cmd, basename, filename):
 
 
 def warn_depends_obsolete(cmd, basename, filename):
-    if os.path.exists(filename):
-        EggInfoDeprecationWarning.emit(
-            "Deprecated config.",
-            """
-            'depends.txt' is not used by setuptools >= 0.6!
-            Configure your dependencies via `setup.cfg` or `pyproject.toml` instead.
-            """,
-            see_docs="userguide/declarative_config.html",
-            due_date=(2023, 6, 1),
-            # Old warning, introduced in 2005, it might be safe to remove soon.
-        )
+    """
+    Unused: left to avoid errors when updating (from source) from <= 67.8.
+    Old installations have a .dist-info directory with the entry-point
+    ``depends.txt = setuptools.command.egg_info:warn_depends_obsolete``.
+    This may trigger errors when running the first egg_info in build_meta.
+    TODO: Remove this function in a version sufficiently > 68.
+    """
 
 
 def _write_requirements(stream, reqs):
@@ -771,26 +745,6 @@ def write_entries(cmd, basename, filename):
     eps = _entry_points.load(cmd.distribution.entry_points)
     defn = _entry_points.render(eps)
     cmd.write_or_delete_file('entry points', filename, defn, True)
-
-
-def get_pkg_info_revision():
-    """
-    Get a -r### off of PKG-INFO Version in case this is an sdist of
-    a subversion revision.
-    """
-    EggInfoDeprecationWarning.emit(
-        "Deprecated API call",
-        "get_pkg_info_revision is deprecated.",
-        due_date=(2023, 6, 1),
-        # Warning introduced in 11 Dec 2015, should be safe to remove
-    )
-    if os.path.exists('PKG-INFO'):
-        with io.open('PKG-INFO') as f:
-            for line in f:
-                match = re.match(r"Version:.*-r(\d+)\s*$", line)
-                if match:
-                    return int(match.group(1))
-    return 0
 
 
 def _egg_basename(egg_name, egg_version, py_version=None, platform=None):
