@@ -2,7 +2,7 @@
 Finalize the repo for a release. Invokes towncrier and bumpversion.
 """
 
-__requires__ = ['bump2version', 'towncrier']
+__requires__ = ['bump2version', 'towncrier', 'jaraco.develop>=7.21']
 
 
 import subprocess
@@ -10,28 +10,14 @@ import pathlib
 import re
 import sys
 
-
-def release_kind():
-    """
-    Determine which release to make based on the files in the
-    changelog.
-    """
-    # use min here as 'major' < 'minor' < 'patch'
-    return min(
-        'major'
-        if 'breaking' in file.name
-        else 'minor'
-        if 'change' in file.name
-        else 'patch'
-        for file in pathlib.Path('changelog.d').iterdir()
-    )
+from jaraco.develop import towncrier
 
 
 bump_version_command = [
     sys.executable,
     '-m',
     'bumpversion',
-    release_kind(),
+    towncrier.release_kind(),
 ]
 
 
@@ -42,16 +28,7 @@ def get_version():
 
 
 def update_changelog():
-    cmd = [
-        sys.executable,
-        '-m',
-        'towncrier',
-        'build',
-        '--version',
-        get_version(),
-        '--yes',
-    ]
-    subprocess.check_call(cmd)
+    towncrier.run('build', '--yes')
     _repair_changelog()
 
 
@@ -78,30 +55,9 @@ def ensure_config():
     subprocess.check_output(['git', 'config', 'user.email'])
 
 
-def check_changes():
-    """
-    Verify that all of the files in changelog.d have the appropriate
-    names.
-    """
-    allowed = 'deprecation', 'breaking', 'change', 'doc', 'misc'
-    except_ = 'README.rst', '.gitignore'
-    news_fragments = (
-        file
-        for file in pathlib.Path('changelog.d').iterdir()
-        if file.name not in except_
-    )
-    unrecognized = [
-        str(file)
-        for file in news_fragments
-        if not any(f".{key}" in file.suffixes for key in allowed)
-    ]
-    if unrecognized:
-        raise ValueError(f"Some news fragments have invalid names: {unrecognized}")
-
-
 if __name__ == '__main__':
     print("Cutting release at", get_version())
     ensure_config()
-    check_changes()
+    towncrier.check_changes()
     update_changelog()
     bump_version()
